Metrics: Unify all Auto-Collection Extractors to ue the same Metric Manager.
