+++ src/Core/Managed/Shared/Extensibility/Implementation/Metrics/SimpleMetricStatisticsAggregator.cs
