+++ src/TelemetryChannels/ServerTelemetryChannel/NuGet/ApplicationInsights.config.install.xdt
@@ -10,12 +10,9 @@
   </TelemetryProcessors>
 
   <TelemetryProcessors xdt:Transform="InsertIfMissing">
-    <Add Type="Microsoft.ApplicationInsights.Extensibility.RequestMetricExtractor, Microsoft.ApplicationInsights"
+    <Add Type="Microsoft.ApplicationInsights.Extensibility.AutocollectedMetricsExtractor, Microsoft.ApplicationInsights"
          xdt:Transform="InsertIfMissing"
          xdt:Locator="Match(Type)"></Add>
-    <Add Type="Microsoft.ApplicationInsights.Extensibility.DependencyMetricExtractor, Microsoft.ApplicationInsights"
-        xdt:Transform="InsertIfMissing"
-        xdt:Locator="Match(Type)"></Add>
     <Add Type="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.AdaptiveSamplingTelemetryProcessor, Microsoft.AI.ServerTelemetryChannel"
          xdt:Transform="InsertIfMissing"
          xdt:Locator="Match(Type)">
